Split test file into two.
